Rename the folder to "generic-components".
