+++ src/lib/UI/domain-components/cards/RecipeCard.svelte
@@ -11,8 +11,8 @@
   import type { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
   import type { Recipe } from "$lib/domain/entities/Recipe";
   import { naming } from "$lib/domain/naming";
-  import Card from "$lib/UI/utility-components/Card.svelte";
-  import FlexRow from "$lib/UI/utility-components/FlexRow.svelte";
+  import Card from "$lib/UI/generic-components/Card.svelte";
+  import FlexRow from "$lib/UI/generic-components/FlexRow.svelte";
 
   // Props:
