+++ src/lib/UI/generic-components/modals/components/CloseButton.svelte
